% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_missing_columns}
\alias{add_missing_columns}
\title{Add columns if missing from input dataframe}
\usage{
add_missing_columns(df, columns, default_value = 0)
}
\arguments{
\item{df}{input data.frame}

\item{columns}{character vector, columns to add}

\item{default_value}{}
}
\value{

}
\description{
Add columns if missing from input dataframe
}
