% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ptemarbs.R
\name{generate_ptemarbs}
\alias{generate_ptemarbs}
\title{Generating PTEMARB files from sources end emissions.}
\usage{
generate_ptemarbs(
  folder,
  sources,
  emissions,
  utc_zone,
  species = list(SO2 = 32, NO = 30, NO2 = 46, PPM25 = 1),
  scenario = "scenario",
  prefix = "ptemarb_",
  begin_date = min(emissions$date),
  end_date = max(emissions$date)
)
}
\arguments{
\item{folder}{}

\item{sources}{}

\item{emissions}{}

\item{utc_zone}{example: UTC+0800}

\item{species}{}

\item{scenario}{}

\item{prefix}{}

\item{begin_date}{}

\item{end_date}{}

\item{utm_zone}{example: 52N}
}
\description{
The script will generate ONE PTEMARB FILE per cluster in indicated folder.
}
