% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_postprocessing.R
\name{get_cp_period}
\alias{get_cp_period}
\title{Retrieve a CALPOST processing period corresponding to a full calendar year.}
\usage{
get_cp_period(params)
}
\description{
Retrieve a CALPOST processing period corresponding to a full calendar year.
}
