% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_srcnam}
\alias{make_srcnam}
\title{Make source names valid for CALPUFF}
\usage{
make_srcnam(x)
}
\arguments{
\item{x}{character vector; names of sources to shorten and make unique}
}
\description{
Make source names valid for CALPUFF
}
