% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_tifs}
\alias{make_tifs}
\title{Title}
\usage{
make_tifs(
  calpuff_files,
  grids,
  ext = "",
  queue = NULL,
  subsets = F,
  max.ranks = 1,
  overwrite = F,
  nmax = 8,
  idp = 1.5,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
