% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_results}
\alias{plot_results}
\title{Title}
\usage{
plot_results(
  calpuff_files,
  scenario_names = calpuff_files$scenarioName,
  dir = dirname(calpuff_files$path[1]),
  map_res = 1,
  plants = NULL,
  plant_names = NULL,
  get_plants = function(x, ...) {     return(x) },
  adm_level = 0,
  plot_km = 400,
  plot_bb = NULL,
  cities = NULL,
  colorkeybasis = NULL,
  queue = seq_along(calpuff_files$path),
  filename_suffix = "",
  outputs = c("png", "kml", "expPop", "cityconcs"),
  zipping_function = zip::zip
)
}
\arguments{
\item{dir}{}

\item{map_res}{in kilometers}
}
\value{

}
\description{
Title
}
