% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ptemarbs.R
\name{ptemarb.create_config_file}
\alias{ptemarb.create_config_file}
\title{Generate a config file for our python PTEMARB generator}
\usage{
ptemarb.create_config_file(
  f_config,
  f_sources,
  f_emissions,
  folder,
  begin_date,
  end_date,
  scenario,
  utc_zone,
  species,
  record_duration_second = 3600,
  default_stack_height_m = 85,
  default_stack_diameter_m = 5,
  prefix = "ptemarb_",
  dataset = "PTEMARB.DAT",
  dataver = "2.1"
)
}
\arguments{
\item{begin_date}{}

\item{end_date}{}

\item{scenario}{}

\item{utc_zone}{}

\item{species}{}

\item{record_duration_second}{}

\item{default_stack_height_m}{}

\item{default_stack_diameter_m}{}

\item{prefix}{}

\item{dataset}{}

\item{dataver}{}

\item{f}{}

\item{file_emission}{}
}
\value{

}
\description{
Generate a config file for our python PTEMARB generator
}
