% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{raster2contourPolys}
\alias{raster2contourPolys}
\title{Create contour polygons from a raster}
\usage{
raster2contourPolys(r, levels = NULL)
}
\arguments{
\item{r}{Input raster}

\item{levels}{Contour levels. Numeric vector, or "auto" to set automatically (default).}
}
\description{
credit: StackOverflow user 'Paul Regular'
}
