% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_calmet.R
\name{runCalmet}
\alias{runCalmet}
\title{Generate input for CALMET}
\usage{
runCalmet(
  input_xls = NULL,
  sources = NULL,
  wrf_dir,
  grid_names = NULL,
  select_grids = "*",
  expand_grids = "*",
  expand_ncells = -5,
  crop_grid = NULL,
  output_dir,
  gis_dir,
  calmet_exe,
  calmet_templates,
  only_make_additional_files = F,
  out_files = NULL,
  boundaries_for_plotting = NULL,
  run_calmet = F
)
}
\arguments{
\item{wrf_dir}{directory where to find m3d files}

\item{expand_grids}{regex string - if grid name matches, the scripts attempts to expand the grid}

\item{expand_ncells}{number of cells to expand the selected grid(s) in each direction (use negative values to crop domain area)}

\item{output_dir}{directory where to write generated input and bat files}

\item{calmet_exe}{filepath (full) of CALMET exe}

\item{calmet_templates}{(list) list with noobs and surfobs template full file paths}

\item{only_make_additional_files}{F: make files anyway and do not skip; T: skip if output file already exists}

\item{out_files}{MYSTERY: this one seems to be generated after this function is called, yet used here}

\item{run_calmet}{whether to actually run CALMET}

\item{gis_folder}{directory where to find land cover files}
}
\description{
Generate input for CALMET
}
