% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_calpuff.R
\name{runCalpuff}
\alias{runCalpuff}
\title{Generate CALPUFF input}
\usage{
runCalpuff(
  run_name,
  params_allgrids,
  emissions_data = NULL,
  source_names = NULL,
  FGD = NULL,
  species_configuration = NULL,
  emitted_polls = list(so2_nox_pm_hg = c("SO2", "NO", "NO2", "PM15", "PM10", "PPM25",
    "HG0", "RGM", "Hgp"), so2_nox_pm = c("SO2", "NO", "NO2", "PM15", "PM10",
    "PPM25"))[[species_configuration]],
  receptors = NULL,
  bgconcs = NULL,
  o3dat = NULL,
  addparams = list(),
  addsubgroups = NULL,
  output_dir = unique(dirname(params_allgrids$METDAT)),
  gis_dir = get_gis_dir(),
  calpuff_exe = "C:/CALPUFF/CALPUFF_v7.2.1_L150618/calpuff_v7.2.1.exe",
  only_make_additional_files = F,
  calpuff_template = list(so2_nox_pm_hg = system.file("extdata",
    "CALPUFF_7.0_template_Hg.INP", package = "creapuff"), so2_nox_pm =
    system.file("extdata", "CALPUFF_7.0_template.INP", package =
    "creapuff"))[[species_configuration]]
)
}
\arguments{
\item{params_allgrids}{}

\item{output_dir}{}

\item{gis_dir}{}

\item{calpuff_exe}{}

\item{only_make_additional_files}{}

\item{calpuff_template}{}

\item{wrf_dir}{}
}
\value{

}
\description{
Generate CALPUFF input
}
